fix Windows build (broken since fa3998b58ec65948e5ac4102cd0e7e0233bb1b91)

+++ test/database.cpp
@@ -250,7 +250,7 @@ TEST_CASE("save points", M) {
   Statement *select = db.prepare("SELECT COUNT(*) FROM test");
 
   auto count = [select] {
-    int count = -255;
+    int64_t count = -255;
     select->exec([&] { count = select->intColumn(0); return false; });
     return count;
   };
